web-stack/backend/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pymongo==4.5.0
python-dotenv==1.0.0